Rename viewport files
